refactor(schemas): split _validators into five responsibility modules
backend/schemas/_validators.py reached 961 lines — over the CI file-size
gate's 900-line hard cap and unlisted, so the gate has failed on main since
roughly PR #129. Split it by responsibility instead of allowlisting:

- _validators_tenant.py: lender-name + state-footprint providers, public
  lender refs, reviewed geography labels
- _validators_protected_class_patterns.py: the reviewed protected-class /
  health / proxy vocabulary regexes, kept alone for auditability
- _validators_protected_class.py: confusable folding, audience-claim
  grammar, windowed proxy matching, fail-closed detectors
- _validators_person_names.py: name-shape detection + reviewed non-person
  phrases
- _validators_unsafe_text.py: prompt-injection, confidential, mechanical-PII
  detectors and the composite contains_unsafe_ai_text

All 36 import sites updated to the new homes; no re-export facade so the
module cannot regrow behind one import path. Patterns verified byte-identical
and detectors behavior-equivalent against the pre-split module before
deletion. Gate, ruff, mypy (235 files), and the unit suite pass; the only
remaining local failure is the documented dotenv-missing deploy dry-run test.
