+++ backend/schemas/growth_agent.py
@@ -8,10 +8,8 @@ from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
 
-from backend.schemas._validators import (
-    contains_protected_class_marketing_text,
-    contains_unsafe_ai_text,
-)
+from backend.schemas._validators_protected_class import contains_protected_class_marketing_text
+from backend.schemas._validators_unsafe_text import contains_unsafe_ai_text
 from backend.schemas.borrower_copy_claims import (
     contains_unsupported_borrower_qualification_claim,
 )
