+++ tests/unit/test_round19_selection_consent_governance.py
@@ -18,7 +18,7 @@ from backend.api.outreach import (
     _assert_final_draft_subject,
 )
 from backend.main import app
-from backend.schemas._validators import contains_protected_class_marketing_text
+from backend.schemas._validators_protected_class import contains_protected_class_marketing_text
 from backend.schemas.borrower_cta_evidence import negative_borrower_cta_evidence
 from backend.schemas.growth_agent import GrowthAgentPromptRunRequest
 from backend.schemas.portfolio_campaign import CampaignRecommendationVariant
